fixing issues with circular dependencies, adding eslint rules
