+++ .eslintrc.json
@@ -11,13 +11,16 @@
 	},
 	"plugins": [
 		"@typescript-eslint",
-		"jest"
+		"jest",
+		"import"
 	],
 	"extends": [
 		"eslint:recommended",
 		"plugin:@typescript-eslint/eslint-recommended",
 		"plugin:@typescript-eslint/recommended",
-		"plugin:jest/recommended"
+		"plugin:jest/recommended",
+		"plugin:import/recommended",
+		"plugin:import/typescript"
 	],
 	"rules": {
 		"@typescript-eslint/explicit-module-boundary-types": "off",
