+++ package.json
@@ -16,7 +16,7 @@
     "start:production": "run-p start:main:production start:worker:production",
     "start:main:production": "run-s db:migrate:prod && node .",
     "start:worker:production": "run-s db:migrate:prod && node src/worker",
-    "clean": "rimraf coverage build tmp src/**/*.js src/**/*.js.map",
+    "clean": "rimraf coverage build tmp src/**/*.js src/**/*.js.map test/**/*.js test/**/*.js.map",
     "env": "ts-node env.ts",
     "build": "tsc -p ./tsconfig.json",
     "build:release": "tsc -p ./tsconfig.release.json",
@@ -105,7 +105,7 @@
     "eslint": "~7.26.0",
     "eslint-config-airbnb-base": "^14.0.0",
     "eslint-config-prettier": "~8.3.0",
-    "eslint-plugin-import": "^2.20.2",
+    "eslint-plugin-import": "^2.24.2",
     "eslint-plugin-jest": "~24.3.6",
     "eslint-plugin-jsdoc": "^16.1.1",
     "eslint-plugin-no-only-tests": "^2.3.1",
