+++ src/api/index.ts
@@ -14,7 +14,7 @@ import uninstall from "../jira/uninstall";
 import { serializeJiraInstallation, serializeSubscription } from "./serializers";
 import getRedisInfo from "../config/redis-info";
 import { elapsedTimeMetrics } from "../config/statsd";
-import { queues } from "../worker/main";
+import { queues } from "../worker/queues";
 import { getLogger } from "../config/logger";
 import { Job, Queue } from "bull";
 import { WhereOptions } from "sequelize";
