+++ src/interfaces/common.ts
@@ -1,0 +1,10 @@
+export enum EnvironmentEnum {
+	test = "test",
+	development = "development",
+	production = "production",
+}
+
+export enum BooleanEnum {
+	true = "true",
+	false = "false",
+}
