+++ src/models/installation.ts
@@ -10,7 +10,6 @@ if (!process.env.STORAGE_SECRET) {
 export const getHashedKey = (clientKey: string): string => {
 	const keyHash = crypto.createHmac("sha256", process.env.STORAGE_SECRET || "");
 	keyHash.update(clientKey);
-
 	return keyHash.digest("hex");
 };
