+++ test/api/index.test.ts
@@ -4,12 +4,12 @@ import express, {NextFunction, Request, Response} from "express";
 import Logger from "bunyan";
 import api from "../../src/api";
 
-describe('api/index', () => {
+describe("api/index", () => {
 	beforeEach(async () => {
 		await Subscription.create({
 			gitHubInstallationId: 123,
-			jiraHost: 'http://blah.com',
-			jiraClientKey: 'myClientKey',
+			jiraHost: "http://blah.com",
+			jiraClientKey: "myClientKey",
 			repoSyncState: {
 				installationId: 123
 			},
@@ -48,14 +48,14 @@ describe('api/index', () => {
 				level: "debug",
 				stream: process.stdout
 			});
-			req.session = { jiraHost: 'http://blah.com' };
+			req.session = { jiraHost: "http://blah.com" };
 			next();
 		});
 		app.use("/api", api);
 		return app;
 	};
 
-	test('GET repoSyncState.json', async () => {
+	test("GET repoSyncState.json", async () => {
 		await supertest(await createApp())
 			.get("/api/123/repoSyncState.json")
 			.set("Authorization", "Bearer xxx")
