+++ test/unit/api/api.test.ts
@@ -7,7 +7,7 @@ import { Installation, Subscription } from "../../../src/models";
 import { mocked } from "ts-jest/utils";
 import { mockModels } from "../../utils/models";
 import api from "../../../src/api";
-import { EnvironmentEnum } from "../../../src/config/env";
+import { EnvironmentEnum } from "../../../src/interfaces/common";
 
 jest.mock("../../../src/models");
